Move FileSink outside of the __init__.py module
